+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 3.0.0
+* Converted to work on ffi (for improved performance, and ios support)
+* Improved isolates to stop ui frame drops
+* Added ImageUtilsIsolate to process camera images in isolate instead of ui thread
+* Updated Camera example
+* Better memory usage and freeing
+* Fixed camera example usage
+
+
 ## 2.0.5
 * fixed dart analyses problems (renamed some variables to follow convention)
 ## 2.0.4